Replaces the hard-coded values
The parental management structure has ratings levels. The size
of the bit mask is constant but magic values were being used
instead of named constants. This commit changes them to named
constants. This is based on a patch submitted by Frederic Marchal,
fmarchal at perso dot be. Thanks for this tidy cleanup.

git-svn-id: svn://svn.mplayerhq.hu/dvdnav/trunk/libdvdread@1221 625a9de0-492d-0410-920c-a7e632af1907