Add --enable-libdvdcss to the autotools install
This commit adds the argument --enable-libdvdcss (off by default). When
used it directly links libdvdcss into libdvdread. This prevents
libdvdread from having to use dlopen() to get at the libdvdcss
functions.

Currently this feature is used when building VLC for mingw32 and darwin.

Patch is part of the videolan archive of patches. The patch was brought
to our attention by Rafael Carre. Thanks to Rafael and the videolan
project!