Improve dlopen support on mingw
Not all mingw build environments have libdl.  And those that do often
only have it in dll form which would have to be distributed with the
application.  So allowing use of the builtin version of dlopen in
dvdread is useful. This patch does the right thing if dlopen is not
found on the system and it adds a configure option to force use of the
builtin dlopen.

It also seems that building dlopen broke at some point.  #includes
pointing to the wrong places.

Patch provided by John Stebbins (stebbins AT jetheadev DOT com), thanks!